+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/implementation/models/SnapshotListResult.java
@@ -13,7 +13,9 @@ import com.azure.json.JsonWriter;
 import java.io.IOException;
 import java.util.List;
 
-/** The result of a snapshot list request. */
+/**
+ * The result of a snapshot list request.
+ */
 @Fluent
 public final class SnapshotListResult implements JsonSerializable<SnapshotListResult> {
     /*
@@ -26,12 +28,15 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
      */
     private String nextLink;
 
-    /** Creates an instance of SnapshotListResult class. */
-    public SnapshotListResult() {}
+    /**
+     * Creates an instance of SnapshotListResult class.
+     */
+    public SnapshotListResult() {
+    }
 
     /**
      * Get the items property: The collection value.
-     *
+     * 
      * @return the items value.
      */
     public List<ConfigurationSnapshot> getItems() {
@@ -40,7 +45,7 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
 
     /**
      * Set the items property: The collection value.
-     *
+     * 
      * @param items the items value to set.
      * @return the SnapshotListResult object itself.
      */
@@ -51,7 +56,7 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
 
     /**
      * Get the nextLink property: The URI that can be used to request the next set of paged results.
-     *
+     * 
      * @return the nextLink value.
      */
     public String getNextLink() {
@@ -60,7 +65,7 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
 
     /**
      * Set the nextLink property: The URI that can be used to request the next set of paged results.
-     *
+     * 
      * @param nextLink the nextLink value to set.
      * @return the SnapshotListResult object itself.
      */
@@ -69,6 +74,9 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
         return this;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
@@ -79,32 +87,31 @@ public final class SnapshotListResult implements JsonSerializable<SnapshotListRe
 
     /**
      * Reads an instance of SnapshotListResult from the JsonReader.
-     *
+     * 
      * @param jsonReader The JsonReader being read.
      * @return An instance of SnapshotListResult if the JsonReader was pointing to an instance of it, or null if it was
-     *     pointing to JSON null.
+     * pointing to JSON null.
      * @throws IOException If an error occurs while reading the SnapshotListResult.
      */
     public static SnapshotListResult fromJson(JsonReader jsonReader) throws IOException {
-        return jsonReader.readObject(
-                reader -> {
-                    SnapshotListResult deserializedSnapshotListResult = new SnapshotListResult();
-                    while (reader.nextToken() != JsonToken.END_OBJECT) {
-                        String fieldName = reader.getFieldName();
-                        reader.nextToken();
+        return jsonReader.readObject(reader -> {
+            SnapshotListResult deserializedSnapshotListResult = new SnapshotListResult();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
 
-                        if ("items".equals(fieldName)) {
-                            List<ConfigurationSnapshot> items =
-                                    reader.readArray(reader1 -> ConfigurationSnapshot.fromJson(reader1));
-                            deserializedSnapshotListResult.items = items;
-                        } else if ("@nextLink".equals(fieldName)) {
-                            deserializedSnapshotListResult.nextLink = reader.getString();
-                        } else {
-                            reader.skipChildren();
-                        }
-                    }
+                if ("items".equals(fieldName)) {
+                    List<ConfigurationSnapshot> items
+                        = reader.readArray(reader1 -> ConfigurationSnapshot.fromJson(reader1));
+                    deserializedSnapshotListResult.items = items;
+                } else if ("@nextLink".equals(fieldName)) {
+                    deserializedSnapshotListResult.nextLink = reader.getString();
+                } else {
+                    reader.skipChildren();
+                }
+            }
 
-                    return deserializedSnapshotListResult;
-                });
+            return deserializedSnapshotListResult;
+        });
     }
 }